% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_banding_function.R
\name{age_banding_function}
\alias{age_banding_function}
\title{Create age cohort based on age in months}
\usage{
age_banding_function(age_months)
}
\arguments{
\item{age_months}{Age measured in months}
}
\value{
Converts age into current Performance Summary Dashboard style age banding
}
\description{
Create age cohort based on age in months
}
\examples{
mileage_banding_function(4567)
mileage_banding_function(140784)



}
