% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aider_data_v1}
\alias{aider_data_v1}
\title{Annonymised and randomised credit application data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1152 rows and 18 columns.}
\usage{
data(aider_data_v1)
}
\description{
Over a thousand records of annonymised and randomised credit application data including
numerical, categorical and date-time data used for testing different functions of the
aider package.
}
\examples{
data(aider_data_v1)
}
\keyword{datasets}
