% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{analyse_missing}
\alias{analyse_missing}
\title{Analyse missing values}
\usage{
analyse_missing(df, case_cutoff = 30, var_cutoff = 20)
}
\arguments{
\item{df}{A data frame}

\item{case_cutoff}{A cut-off to narrow down cases with higher percentage of missing values}

\item{var_cutoff}{A cut-off to narrow down variables with higher percentage of missing values}
}
\description{
This function returns a set of basic summary statistics of missing values in a data frame
}
\examples{
analyse_missing(credit_data)
}
