% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{analyse_predictiveness}
\alias{analyse_predictiveness}
\title{Analyse variables predictiveness}
\usage{
analyse_predictiveness(df, target, cutoff)
}
\arguments{
\item{df}{A a data frame}

\item{target}{Target variable}

\item{cutoff}{Exclude variables that have a correlation higher then a specified threshold}
}
\description{
This function calculates importance of all variables and selects the ones that are
the most predictive and uncorrelated to each other based on a selected correlation cutoff.
This process is performed iteratively from the most to least important variables. Only numerical
attributes are considered.
}
\examples{
data <- recipes::credit_data \%>\%
  first_to_lower()
}
