% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{analyse_transformations}
\alias{analyse_transformations}
\title{Find predictive variable transformations}
\usage{
analyse_transformations(df, target)
}
\arguments{
\item{df}{A data frame}

\item{target}{A target variable}
}
\description{
TBD
}
\examples{
data <- credit_data \%>\%
  first_to_lower()

transformations <- analyse_transformations(data, status)
}
