% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{apply_mov}
\alias{apply_mov}
\title{Apply multivariate outlier detection}
\usage{
apply_mov(df)
}
\arguments{
\item{df}{A data frame}
}
\description{
This function calculates individual observations outlier score with the Lof algorithm and
returns the original data frame with the Lof outlier score, Lof stats and variables with the
highest impact on the Lof score, as well as a data frame with the most outlying cases (approx. top 5%).
}
\examples{
data <- recipes::credit_data \%>\%
  apply_recipe(Status) \%>\%
  prep(retain = TRUE) \%>\%
  juice() \%>\%
  select(-Status)

out <- apply_mov(data)
}
