% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{apply_recipe}
\alias{apply_recipe}
\title{Apply a model recipe}
\usage{
apply_recipe(df, target)
}
\arguments{
\item{df}{A data frame}

\item{target}{A target variable}
}
\description{
This function calculates a simple and universal model matrix recipe that can be used
for a number of standardised modelling tasks.
}
\examples{
apply_recipe(recipes::credit_data, Status)
}
