% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{apply_recipe_bp2}
\alias{apply_recipe_bp2}
\title{Apply a bp2 model recipe}
\usage{
apply_recipe_bp2(df, target)
}
\arguments{
\item{df}{A data frame}

\item{target}{A target variable}
}
\description{
This function calculates a bp2 model matrix recipe.
}
\examples{
apply_recipe_bp2(credit_data, Status)
}
