% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{apply_recipe_bp3}
\alias{apply_recipe_bp3}
\title{Apply a bp3 model recipe}
\usage{
apply_recipe_bp3(df, target)
}
\arguments{
\item{df}{A data frame}

\item{target}{A target variable}
}
\description{
This function calculates a bp3 model matrix recipe.
}
\examples{
apply_recipe_bp3(credit_data, Status)
}
