% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{assess_performance}
\alias{assess_performance}
\title{Assess model performance}
\usage{
assess_performance(df, actual = actual, prediction = prediction)
}
\arguments{
\item{df}{A data frame}

\item{actual}{A variable with actual outcome}

\item{prediction}{A variable with outcome prediction}
}
\description{
This function calculates the most importance classification task performance measures.
}
\examples{
assess_performance(credit_data, tbd)
}
