% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{build_predictive_models}
\alias{build_predictive_models}
\title{Build various predictive models}
\usage{
build_predictive_models(df, target, folds = 5, repeats = 1,
  upsample = "no")
}
\arguments{
\item{df}{A data frame}

\item{target}{A target variable}

\item{folds}{Specify the number of folds in cross-validation. Defaults to 5}

\item{repeats}{Specify the number of times the fitting process should be repeated. Defaults to 1}

\item{upsample}{Should the minority class be upsampled during resampling? Defaults to "no"}
}
\description{
This function specifies different predictive models on the dataset. Data preprocessing happens automatically through applying the BP2 recipe blueprint.
}
\examples{
data <- credit_data \%>\%
  first_to_lower()

int <- build_predictive_models(data, status, upsample = "yes")
}
