% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_bad_rate}
\alias{calculate_bad_rate}
\title{Calculate bad rate}
\usage{
calculate_bad_rate(df, performance, top_level = "1")
}
\arguments{
\item{df}{A data drame}

\item{performance}{Performance variable}

\item{top_level}{Top level of the performance variable. Defaults to 1}
}
\description{
This function calculates bad rates for any number of grouping variables.
}
\examples{
data("aider_data_v1")

aider_data_v1 \%>\%
  calculate_bad_rate(target)

aider_data_v1 \%>\%
  mutate(target = ifelse(target == 1, "Npl", "Pl")) \%>\%
  calculate_bad_rate(target, "Npl")

aider_data_v1 \%>\%
  calculate_bad_rate(performance = target, top_level = "1", country)
}
\keyword{risk}
