% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_correlation}
\alias{calculate_correlation}
\title{Calculate tidy correlation}
\usage{
calculate_correlation(df, cutoff = 0.5, method = "pearson",
  use = "pairwise.complete.obs")
}
\arguments{
\item{df}{A a data frame}

\item{cutoff}{Include correlation higher then a threshold. Defaults to 0.50}

\item{method}{Which correlation should be computed. Defaults to pearson}

\item{use}{Which method for computing correlation in presence of missing values. Defaults to pairwise.complete.obs}
}
\description{
This function calculates correlation among all numerical attributes in a tidy format. Only
}
\examples{
data("aider_data_v1")

calculate_correlation(aider_data_v1)
}
\keyword{risk}
