% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_decile_table}
\alias{calculate_decile_table}
\title{Calculate a decile breakdown}
\usage{
calculate_decile_table(df, binning, grouping, top_level = "1", n_bins = 10)
}
\arguments{
\item{df}{A data frame}

\item{binning}{Variable for which binning should be applied}

\item{grouping}{A two-level (binary) variable to calculate the ratio in each bin}

\item{top_level}{Top level of the grouping variable. Defaults to 1}

\item{n_bins}{Provide a number of bins. Defaults to 10}
}
\description{
This function calculates a decile table for any combination of numerical and categorical variables.
}
\examples{
data("aider_data_v1")

aider_data_v1 \%>\%
calculate_decile_table(binning = sc_v3,
                      grouping = target)

aider_data_v1 \%>\%
  select(country, target, sc_v3) \%>\%
  nest(-country) \%>\%
  mutate(stats = map(data, calculate_decile_table, sc_v3, target)) \%>\%
  select(country, stats) \%>\%
  unnest()
}
\keyword{risk}
