% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_importance}
\alias{calculate_importance}
\title{Find the most important variables}
\usage{
calculate_importance(df, target, type = "classification", cutoff = 10)
}
\arguments{
\item{df}{A a data frame}

\item{target}{Target variable}

\item{type}{Type of modelling task. Defaults to classification}

\item{cutoff}{Include only the most predictive variables. Defaults to 10}
}
\description{
This is a function that enables calculating variable importance for numerical attributes.
}
\examples{
data("aider_data_v1")

aider_data_v1 \%>\%
  calculate_importance(target)

aider_data_v1 \%>\%
  calculate_importance(target, type = "regression")
}
\keyword{risk}
