% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_logodds_table}
\alias{calculate_logodds_table}
\title{Calculate a log-odds table}
\usage{
calculate_logodds_table(df, binning, grouping, top_level = "1", n_bins = 10,
  risk_names = TRUE, format = FALSE)
}
\arguments{
\item{df}{A data frame}

\item{binning}{Variable for which binning should be applied}

\item{grouping}{A two-level (binary) variable to calculate the ratio in each bin}

\item{top_level}{Top level of the grouping variable. Defaults to 1}

\item{n_bins}{Provide a number of bins. Defaults to 10}

\item{risk_names}{Should column names be converted to risk-specific names? Defaults to TRUE}

\item{format}{Should table printing be formatted with kable? Defaults to FALSE}
}
\description{
This function calculates a log-odds table for any combination of numerical and categorical variables.
}
\examples{
credit_data \%>\%
  first_to_lower() \%>\%
  calculate_logodds_table(binning = time,
                          grouping = status,
                          top_level = "bad")

credit_data \%>\%
  first_to_lower() \%>\%
  select(marital, status, time) \%>\%
  nest(-marital) \%>\%
  mutate(stats = map(data, calculate_logodds_table, time, status, "bad")) \%>\%
  select(marital, stats) \%>\%
  unnest()
}
