% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_share}
\alias{calculate_share}
\title{Calcualte chare}
\usage{
calculate_share(df, grouping)
}
\arguments{
\item{df}{A a data frame}

\item{grouping}{Variable to calculate the share for}
}
\description{
This function calculates the share of grand total for any number of grouping variables.
}
\examples{
data("aider_data_v1")

aider_data_v1 \%>\%
  calculate_share(country)

aider_data_v1 \%>\%
  nest(-country) \%>\%
  mutate(stats = map(data, calculate_share, industry)) \%>\%
  select(country, stats) \%>\%
  unnest()
}
\keyword{risk}
