% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_stats_numeric}
\alias{calculate_stats_numeric}
\title{Calculate statistics}
\usage{
calculate_stats_numeric(df)
}
\arguments{
\item{df}{A data frame}
}
\description{
This function calculates statistics for numerical attributes for any number of groupong variables.
}
\examples{
data("aider_data_v1")

aider_data_v1 \%>\%
  nest(-country) \%>\%
  mutate(stats = map(data, ~calculate_stats_numeric(.x))) \%>\%
  select(country, stats) \%>\%
  unnest()
}
\keyword{risk}
