% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{calibrate_probabilities}
\alias{calibrate_probabilities}
\title{This function calibrates model predicted probabilities by aligning them with the actual outcome variable through minimizing
the log-loss with Platt scaling. Calibrating to a different event rate is also possible with the 'ct' parameter.}
\usage{
calibrate_probabilities(df_pred, target, prediction, top_level = "1",
  ct = 0)
}
\arguments{
\item{df_pred}{A data frame containing raw model predictions and the target even}

\item{target}{Actual outcome variable}

\item{prediction}{Predicted outcome variable}

\item{top_level}{Top level of the grouping variable. Defaults to "1"}

\item{ct}{Central tendency/event rate to which the prediction is to be calibrated. Defaults to 0 if only the log-loss should be optimized}
}
\description{
This function calibrates model predicted probabilities by aligning them with the actual outcome variable through minimizing
the log-loss with Platt scaling. Calibrating to a different event rate is also possible with the 'ct' parameter.
}
\examples{
set.seed(42)

outcome <- data_frame(
      default = rbinom(100, 1, 0.5),
      pred = runif(100, 0, 1)
        ) \%>\%
      calibrate_probabilities(default, pred, "1")
}
