% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak.R
\name{cap_at_percentile}
\alias{cap_at_percentile}
\title{Cap numeric values at selected percentiles}
\usage{
cap_at_percentile(x, floor = 0.025, roof = 0.975)
}
\arguments{
\item{x}{A vector or data frame column}

\item{floor}{A bottom percentile. Defaults to 0.975}

\item{roof}{A top percentile. Defaults to 0.025}
}
\description{
This function capps values of a numeric vector at a specified percentile.
Can be used both with numeric vectors of data frames with mutate and map_at.
}
\examples{
x <- seq(1, 100, 1)
cap_at_percentile(x)

data <- data.frame(x = seq(1, 100, 1))
data \%<>\%
  mutate(y = cap_at_percentile(x))

}
\keyword{cap}
