% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak.R
\name{cap_between}
\alias{cap_between}
\title{Cap numeric values between two values}
\usage{
cap_between(x, floor = NA, roof = NA)
}
\arguments{
\item{x}{A vector or data frame column}

\item{floor}{A bottom number. Defaults to NA}

\item{roof}{A top number. Defaults to NA}
}
\description{
This function capps values of a numeric vector between specified values.
Can be used both with numeric vectors of data frames with mutate and map_at.
}
\examples{
x <- seq(1, 100, 1)
cap_between(x, 40, 60)

data <- data.frame(x = seq(1, 100, 1))

data \%<>\%
  mutate(y = cap_between(x, 40, 60))

}
\keyword{cap}
