% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak.R
\name{change_names}
\alias{change_names}
\title{Change a vector or data frame names}
\usage{
change_names(x, from = ".", to = "_")
}
\arguments{
\item{x}{A data frame or vector}

\item{from}{Character to replace. Defaults to "."}

\item{to}{Character to apply Defaults to "_"}
}
\description{
This function changes your data frame column names or character vector elements according with a specified pattern.
By default it fixes column names that have a "." used as words separator to "_".
}
\examples{
x <- c("Example.1", "Example.2", "Example/3")
change_names(x)

df <- data_frame(
  "Example.1" = c(1, 2),
  "Example.2" = c(1, 2),
  "Example!3" = c(1, 2))

df \%>\%
  change_names()
}
