% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{compare_samples}
\alias{compare_samples}
\title{Compare samples similarity}
\usage{
compare_samples(df)
}
\arguments{
\item{df}{A a data frame}
}
\description{
This function estimated a model to compare two samples similarity and the most predictive variables.
}
\examples{
data <- credit_data \%>\%
  first_to_lower()

data_v1 <- data \%>\%
  select(-status) \%>\%
  filter(records \%in\% c("yes")) \%>\%
  mutate(status = 1)

data_v2 <- data \%>\%
  select(-status) \%>\%
  filter(records \%in\% c("no")) \%>\%
  mutate(status = 0)

data_cmb <- bind_rows(
  data_v1,
  data_v2
  ) \%>\%
  mutate(status = ifelse(status == 1, "yes", "no")) \%>\%
  rename(target = status) \%>\%
  compare_samples()
}
