% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_date.R
\name{convert_to_date}
\alias{convert_to_date}
\title{Convert string into date}
\usage{
convert_to_date(date)
}
\arguments{
\item{date}{Date object. Must be in YYYY-mm-dd format}
}
\value{
A date object in ISO 8601 international standard
}
\description{
Convert string into date
}
\examples{
convert_to_date("2022-06-17")

}
