% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{count_proportions}
\alias{count_proportions}
\title{Count proportions of levels}
\usage{
count_proportions(x)
}
\arguments{
\item{x}{A vector}
}
\description{
This function calculates the proportions of unique observations in a vector.
}
\examples{
count_proportions(credit_data$Marital)
}
