% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_colours.R
\name{cox_colours}
\alias{cox_colours}
\title{Loads Hexidecimal codes for Cox Automotive colour palette}
\usage{
cox_colours()
}
\value{
Object containing hexidecimal for each colour. Objects are called cax_grey, cox_dark_blue, cox_aqua, cox_teal, cix_sky_blue, cox_berry, cox_fire, cox_lime, cox_dark_grey
}
\description{
Creates an object containing the hexidecimal code for the following colours: Grey, dark blue, aqua, teal, sky blue, berry, fire, lime, dark grey
}
\examples{
cox_colours()

}
