% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rodbc.R
\name{extract_rodbc}
\alias{extract_rodbc}
\title{Extract data from Database using RODBC}
\usage{
extract_rodbc(sql, odbc_connection = "databricks_connection")
}
\arguments{
\item{sql}{SQL query}

\item{odbc_connection}{name of your ODBC connection. Defaults to the authors ODBC connector}
}
\description{
Uses the RODBC package to extract data from a  database using an ODBC connector. Sets
}
\examples{
\dontrun{sql_query <- "SELECT * FROM database.table"
extract_rodbc(sql_query)}



}
