% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{find_interactions}
\alias{find_interactions}
\title{Find meaningful variable interactions}
\usage{
find_interactions(df, target)
}
\arguments{
\item{df}{A data frame}

\item{target}{A target variable}
}
\description{
This function calculates a set of potentially meaningful side-effects. Second level interactions are calculated
through estimating and tuning two models with repeated cross-validation: MARS and Lasso regression.
Data preprocessing happens automatically through applying the BP2 recipe blueprint.
}
\examples{
data <- credit_data \%>\%
  first_to_lower()

int <- find_interactions(data, status)
}
