% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{find_most_predictive}
\alias{find_most_predictive}
\title{Find most predictive variables}
\usage{
find_most_predictive(df, target, cutoff)
}
\arguments{
\item{df}{A a data frame}

\item{target}{Target variable}

\item{cutoff}{Exclude variables that have a correlation higher then a specified threshold}
}
\description{
This function calculates importance of all variables and selects the ones that are
the most predictive and uncorrelated to each other based on a selected correlation cutoff.
This process is performed iteratively from the most to least important variables. Only numerical
attributes are considered.
}
\examples{
data <- credit_data \%>\%
  first_to_lower()

find_most_predictive(data, status, 0.5)
}
