% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak.R
\name{first_to_lower}
\alias{first_to_lower}
\title{Convert strings beginning to lowercase}
\usage{
first_to_lower(x)
}
\arguments{
\item{x}{A data frame or vector}
}
\description{
This function converts your data frame column names of vector elements first letter to lowercase.
}
\examples{
data("aider_data_v1")

aider_data_v1 \%<>\%
  first_to_lower()
}
\keyword{names}
