% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak.R
\name{first_to_upper}
\alias{first_to_upper}
\title{Convert strings beginning to uppercase}
\usage{
first_to_upper(x)
}
\arguments{
\item{x}{A data frame or vector}
}
\description{
This function converts your data frame column names of vector elements first letter to uppercase.
}
\examples{
data("aider_data_v1")

aider_data_v1 \%<>\%
  first_to_lower()
}
\keyword{names}
