% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak.R
\name{form_name}
\alias{form_name}
\title{Svae files with consistent naming across your sessions}
\usage{
form_name(name, path, body, date, ext)
}
\arguments{
\item{path}{A path (folder/) withing your project structure}

\item{body}{Body part of the file name}

\item{ext}{Extension of the file. Should be used consistently with the saving function}
}
\description{
This function sets consistent names for files to be saved.
}
\examples{
name <- "Spotcap_LightPerformanceModel_6M"
date <- Sys.Date()
read_csv(form_name("03 ModelBuildingFiles", "Outliers", "csv"))
}
