% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak.R
\name{format_my_table}
\alias{format_my_table}
\title{Format a table nicely}
\usage{
format_my_table(df, width = TRUE)
}
\arguments{
\item{df}{A data frame}

\item{width}{Should the table have full-page width. Defaults to TRUE}
}
\description{
This function creates nicely formatted tables in R markdown. It also works with the 'formattable' package.
}
\examples{
data("aider_data_v1")

aider_data_v1 \%>\%
  calculate_share(country) \%>\%
  format_my_table()
}
\keyword{table}
