% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{get_aider_palletes}
\alias{get_aider_palletes}
\title{Use aider colors schemes}
\usage{
get_aider_palletes()
}
\description{
This function creates different pallete vectors to be used for ggplots.
}
\examples{
get_aider_palletes()
}
\keyword{pallete}
