% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqr_.R
\name{iqr_}
\alias{iqr_}
\title{The Interquartile Range excluding NA values}
\usage{
iqr_(x)
}
\arguments{
\item{x}{numeric vector whose sample quantiles are wanted}
}
\description{
Tweak to base R IQR function. Calculates the interquartile range of vector with NAs removed
}
\examples{
iqr_(c(9, 12, 376, 483, NA))
}
