% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_dir.R
\name{make_data_dir}
\alias{make_data_dir}
\title{Create data directory for project}
\usage{
make_data_dir(
  project_path = dirname(rstudioapi::getActiveProject()),
  project_name = stringr::word(project_path, 3, sep = "/"),
  data_path_root = "../../data"
)
}
\arguments{
\item{project_path}{file path of your R project. Defaults to automatically extract the information using the rstudioAPI}

\item{project_name}{name of the data directory. Defaults to use the same name as your R project. Will extract it from project_path if project is second sub-directory from system root. e.g. 'C:/git/my_project'}

\item{data_path_root}{file path of your data directory Defaults to '../../data'}
}
\description{
Creates a directory with the same name as your R project. Defaults to write at '../../data'
}
\examples{

\dontrun{
make_data_dir()
make_data_dir(project_name = "my_project", data_path_root = "../../../data")}



}
