% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_.R
\name{max_}
\alias{max_}
\title{Maximum of vector elements excluding NA values}
\usage{
max_(x)
}
\arguments{
\item{x}{An R object. Numeric/logical vector and date/date-time/time interval objects}
}
\description{
Tweak to base R max function. Calculates maximum of vector excluding NAs
}
\examples{
max_(c(9, 12, 376, 483, NA))

}
