% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_.R
\name{mean_}
\alias{mean_}
\title{Mean of vector elements excluding NA values}
\usage{
mean_(x)
}
\arguments{
\item{x}{An R object. Numeric/logical vector and date/date-time/time interval objects}
}
\description{
Tweak to base R mean function. Calculates mean of vector excluding NAs
}
\examples{
mean_(c(9, 12, 376, 483, NA))

}
