% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_.R
\name{median_}
\alias{median_}
\title{Median of vector elements excluding NA values}
\usage{
median_(x)
}
\arguments{
\item{x}{An R object. Numeric/logical vector and date/date-time/time interval objects}
}
\description{
Tweak to base R median function. Calculates median of vector excluding NAs
}
\examples{
median_(c(9, 12, 376, 483, NA))

}
