% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mileage_banding_function.R
\name{mileage_banding_function}
\alias{mileage_banding_function}
\title{Create mileage cohort based on age in months}
\usage{
mileage_banding_function(mileage)
}
\arguments{
\item{mileage}{The mileage of the vehicle}
}
\description{
Create mileage cohort based on age in months
}
\details{
Converts mileage into current Performance Summary Dashboard style mileage banding
}
\examples{
mileage_banding_function(4567)
mileage_banding_function(140784)



}
