% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_.R
\name{min_}
\alias{min_}
\title{Minimum of vector elements excluding NA values}
\usage{
min_(x)
}
\arguments{
\item{x}{An R object. Numeric/logical vector and date/date-time/time interval objects}
}
\description{
Tweak to base R min function. Calculates minimum of vector excluding NAs
}
\examples{
min_(c(9, 12, 376, 483, NA))
}
