% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_count.R
\name{na_count}
\alias{na_count}
\title{Count number of NA values within a data frame}
\usage{
na_count(x)
}
\arguments{
\item{x}{data frame or array of interest}
}
\description{
Counts the number of NAs within a data frame or array
}
\examples{
\dontrun{na_count(iris)}


}
