% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{number_ticks}
\alias{number_ticks}
\title{Equal sized axis}
\usage{
number_ticks(n = 10)
}
\arguments{
\item{n}{Number of ticks}
}
\description{
This function created equally spaced axis ticks for ggplot graphs.
}
\keyword{ticks}
