% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_theme.R
\name{output_theme}
\alias{output_theme}
\title{Change default ggplot2 theme}
\usage{
output_theme(title_font_size = 14, axis_font_size = 12)
}
\arguments{
\item{title_font_size}{Font size of plot title. Defaults to 14}

\item{axis_font_size}{Font size of axis labels. Defaults to 12}
}
\value{
Object containing hexidecimal for each colour. Objects are called cax_grey, cox_dark_blue, cox_aqua, cox_teal, cix_sky_blue, cox_berry, cox_fire, cox_lime, cox_dark_grey
}
\description{
Creates default ggplot output to Data Science ggplot2 output theme
}
\examples{
output_theme()
output_theme(18, 14)

}
