% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_bars}
\alias{plot_bars}
\title{Plot bar-plots of numerical variables}
\usage{
plot_bars(df, x, fill = NULL, facet = c(" "), angle = 0,
  title = TRUE, binwidth = NULL, position = "dodge",
  lab_x = "Value range", lab_y = "Count", legend = TRUE,
  vline = c(Inf), alpha = 1, quantile_low = 0.025,
  quantile_high = 0.975, palette = "cartography")
}
\arguments{
\item{df}{A data frame}

\item{x}{A categorical variable for the x axis groupings}

\item{fill}{Select an additional grouping variable to be used for plotting. Defaults to NULL}

\item{facet}{Select an additional faceting variable to create facets. Defaults to c(" ")}

\item{angle}{Select the rotation angle for the x axis labels. Defaults to 0}

\item{title}{Should the plot title appear automatically. Defaults to TRUE}

\item{binwidth}{Select binwidth, defaults to NULL}

\item{position}{Select the position of the barplot. Defaults to "dodge",}

\item{lab_x}{Text that is displayed on the x axis. Defaults to "Level"}

\item{lab_y}{Text that is displayed on the y axis. Defaults to "Value range"}

\item{legend}{Should the plot legend appear automatically. Defaults to TRUE}

\item{vline}{Should any horizontal lines be added to the plot. Defaults to c(Inf)}

\item{alpha}{Select plot fill transparency. Defaults to 1}

\item{quantile_low}{Select lower percentile for outliers exclusion. Defaults to 2.5\%}

\item{quantile_high}{Select upper percentile for outliers exclusion. Defaults to 97.5\%}

\item{palette}{Select a color palette from colors available in the select_palette function}
}
\description{
This function creates nicely formatted, standardised bar-plots.
}
\examples{

data <- credit_data \%>\%
 first_to_lower()

data \%>\%
 plot_bars(x = income)

data \%>\%
 plot_bars(x = income,
           fill = marital)
data \%>\%
 plot_bars(x = income,
           fill = marital,
           facet = job)
data \%>\%
 plot_bars(x = income,
           fill = marital,
           facet = job,
           position = "stack",
           binwidth = 50,
           vline = 45,
           angle = 45,
           alpha = .7,
           palette = "berlin")


}
