% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Plot a correlation matrix of numerical variables}
\usage{
plot_correlation(df)
}
\arguments{
\item{df}{A data frame}

\item{method}{A character string indicating which correlation coefficient (or covariance) is to be computed. One of "spearman" (default), "pearson" or "kendall": can be abbreviated}
}
\description{
This function creates a nicely formatted, standardised correlation matrix of numerical variables. Long variables names should be shortened before for easier interpretation.
}
\examples{
credit_data \%>\% plot_correlation()
}
