% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_density}
\alias{plot_density}
\title{Create a density plot}
\usage{
plot_density(df, x, fill = NULL, facet = c(" "), ticks = 10, angle = 0,
  title = TRUE, legend = TRUE, vline = c(Inf), alpha = 0.7,
  quantile_low = 0.025, quantile_high = 0.975)
}
\arguments{
\item{df}{A data frame}

\item{x}{A numerical variable to plot its density}

\item{fill}{Select an additional grouping variable to be used for density plotting. Defaults to NULL}

\item{facet}{Select an additional faceting variables to create facets. Defaults to c(" ")}

\item{ticks}{Select the number of ticks on the x and y axis. Defaults to 10}

\item{angle}{Select the rotation angle for the x axis labels. Defaults to 0}

\item{title}{Should the plot title appear automatically. Defaults to TRUE}

\item{legend}{Should the plot legend appear automatically. Defaults to TRUE}

\item{vline}{Should any vertical lines be added to the plot. Defaults to c(Inf)}

\item{alpha}{Select plot transparency. Defaults to .5}

\item{quantile_low}{Select lower percentile for outliers exclusion. Defaults to 2.5%}

\item{quantile_high}{Select upper percentile for outliers exclusion. Defaults to 97.5%}
}
\description{
This function creates nicely formatted, standardised density plots.
}
\examples{
data("aider_data_v1")

aider_data_v1 \%>\%
  plot_density(x = in_sales_v1)

aider_data_v1 \%>\%
  plot_density(x = in_sales_v1,
               facet = country)

aider_data_v1 \%>\%
  plot_density(x = in_sales_v1,
               fill = country,
               facet = country,
               ticks = 10,
               title = TRUE,
               vline = 1500000,
               legend = TRUE,
               alpha = .5)
}
\keyword{plot}
