% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_logodds}
\alias{plot_logodds}
\title{Plot a log-odds table}
\usage{
plot_logodds(df, title = "Evaluation of log-odds linearity",
  lab_x = "Mean of variable deciles", lab_y = "Log-odds")
}
\arguments{
\item{df}{A data frame}

\item{title}{Text that is displayed on as the plot title. Defaults to "Lift chart: evaluation of model predicted probabilities vs. actual defaul rates across deciles"}

\item{lab_x}{Text that is displayed on the x axis. Defaults to "Mean of variable deciles"}

\item{lab_y}{Text that is displayed on the y axis. Defaults to "Log-odds"}
}
\description{
This function creates a nicely formatted, standardised log-odds plot. Prior to calling the function
the data should only be in a form of a log-odds table (calculate_logodds_table() function will
do that for you), unless it's already provided.
}
\examples{
credit_data \%>\%
  first_to_lower() \%>\%
  calculate_logodds_table(binning = time,
                          grouping = status,
                          top_level = "bad") \%>\%
  plot_logodds()
}
