% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation_titles.R
\name{presentation_titles}
\alias{presentation_titles}
\title{Change variable names to presentation standard}
\usage{
presentation_titles(data)
}
\arguments{
\item{data}{Data frame you want to change the titles}
}
\description{
Change string casing to proper and replace underscrores with spaces
}
\examples{
\dontrun{
presentation_titles(my_data)
my_data \%>\% presentation_titles()}

}
