% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_banding_function.R
\name{price_banding_function}
\alias{price_banding_function}
\title{Create price cohort based on value}
\usage{
price_banding_function(sold_price)
}
\arguments{
\item{sold_price}{The sold price of the vehicle}
}
\description{
Converts sold price into current Performance Summary Dashboard style price banding
}
\examples{
price_banding_function(4567)
price_banding_function(140784)



}
