% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_.R
\name{quantile_}
\alias{quantile_}
\title{Sample quantities excluding NA values}
\usage{
quantile_(x, probs)
}
\arguments{
\item{x}{numeric vector whose sample quantiles are wanted}

\item{probs}{Numeric vector of probabilities with values in between 0 and 1. Defaults to quartiles}
}
\description{
Tweak to base R quantile function. Calculates the sample quantile corresponding to the given probabilities with NAs removed
}
\examples{
quantile_(c(9, 12, 376, 483, NA), probs = seq(0, 1, 0.25))

}
