% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak.R
\name{round_to}
\alias{round_to}
\title{Round values to integers}
\usage{
round_to(x, to = 1000)
}
\arguments{
\item{x}{A vector or data frame column}

\item{to}{An integer to round to. Defaults to a 1,000}
}
\description{
This function rounds values of a numeric vector to a selected integer, for example: 1,000 or 10,000.
This allows for a more general interpretation of high values.
}
\examples{
round_to(12456)
}
