% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_jpeg.R
\name{save_jpeg}
\alias{save_jpeg}
\title{Export visualisation as JPEG}
\usage{
save_jpeg(path)
}
\arguments{
\item{path}{File path and filename for visualisation WITHOUT the file extension}
}
\value{
A JPEG file with dpi = 300 and height = 8in and width = 13in
}
\description{
Wrapper around ggsave function. Run this function directly after you've created your visualisation
}
\examples{
\dontrun{
save_jpeg('output/myplot')
}

}
