% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_object.R
\name{save_object}
\alias{save_object}
\title{Save object in data directory}
\usage{
save_object(
  x,
  filename,
  project_path = dirname(rstudioapi::getActiveProject()),
  project_name = stringr::word(project_path, 3, sep = "/"),
  data_path_root = "../../data"
)
}
\arguments{
\item{x}{object to save}

\item{filename}{name for saved object. e.g. 'mydata.rda' or 'mydata.rds'}

\item{project_path}{file path of your R project. Defaults to automatically extract the information using the rstudioAPI}

\item{project_name}{name of the data directory. Defaults to use the same name as your R project. Will extract it from project_path if project is second sub-directory from system root. e.g. 'C:/git/my_project'}

\item{data_path_root}{file path of your data directory Defaults to '../../data'}
}
\description{
Uses base R save() function to automatically save to data folder. Use make_data_dir() to configure
}
\examples{
\dontrun{
save_object(iris, "mydata.rda")
save_object(iris, "mydata.rda", data_path_root = '../../../data')}
}
