% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_svg.R
\name{save_svg}
\alias{save_svg}
\title{Export visualisation as SVG}
\usage{
save_svg(path)
}
\arguments{
\item{path}{File path and filename for visualisation WITHOUT the file extension}
}
\value{
A SVG file with dpi = 300 and height = 8in and width = 13in
}
\description{
Wrapper around ggsave function. Run this function directly after you've created your visualisation
}
\examples{
\dontrun{
save_svg('output/myplot')
}

}
