% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{select_palette}
\alias{select_palette}
\title{Palettes are based on the list of available color schemes: https://github.com/EmilHvitfeldt/r-color-palettes. We selected a shortlist of the most sensible palettes for you.}
\usage{
select_palette(palette = "cartography")
}
\arguments{
\item{palette}{Select a palette. Available options are: use "risk" for approved/ rejected, performing/ non-performing palletes, use "cartography" to get
20 discrete colors or "awtools" to get 8 discrete colors, and finally use "berlin" or "lajolla" to get 100 continuous colors. Defaults to "cartography"}
}
\description{
Palettes are based on the list of available color schemes: https://github.com/EmilHvitfeldt/r-color-palettes. We selected a shortlist of the most sensible palettes for you.
}
