% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{set_me_up}
\alias{set_me_up}
\title{Basic R setup}
\usage{
set_me_up()
}
\description{
This function sets up basic R parameters: seed, scipien, R tools path and loads the tidyverse library.
}
\examples{
set_me_up()
}
\keyword{setup}
