% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_.R
\name{sum_}
\alias{sum_}
\title{Sum of vector elements excluding NA values}
\usage{
sum_(x)
}
\arguments{
\item{x}{An R object. Numeric/logical vector and date/date-time/time interval objects}
}
\description{
Tweak to base R sum function. Calculates sum of vector excluding NAs
}
\examples{
sum_(c(9, 12, 376, 483, NA))
}
