% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_table.R
\name{write_table}
\alias{write_table}
\title{Output data}
\usage{
write_table(x, file)
}
\arguments{
\item{x}{data frame}

\item{file}{filepath}
}
\description{
Writes data frame to csv without row names, character or factor columns are not surrounded by quotes, and NAs displayed as enpty strings
}
\examples{
\dontrun{
write_table(iris, "my_directory/my_file.csv")}
}
