% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{aider_theme}
\alias{aider_theme}
\title{Aider ggplot2 theme}
\usage{
aider_theme(type = "grey")
}
\arguments{
\item{type}{Select a theme type. Defaults to "grey", another option includes also "ipsum". Otherwise no theme is applied}
}
\description{
This function applies the aider theme to any ggplot graph in order to
create more complete and nicer looking visualizations.
}
