% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{aider_theme}
\alias{aider_theme}
\title{Aider ggplot2 theme}
\usage{
aider_theme()
}
\description{
This function applies the aider package theme to any ggplot graph. It is based on theme_grey.
}
\keyword{theme}
