% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{apply_rfe}
\alias{apply_rfe}
\title{Apply recursive feature elimination}
\usage{
apply_rfe(df, target, subsets = NULL)
}
\arguments{
\item{df}{A a data frame}

\item{target}{Target variable}

\item{subsets}{Provide a vector of the number of variables to fit models to. Defaults to NULL}
}
\description{
This function performes Recursive Feature Elimination (RFE) based on the implementation in the caret package.
}
\examples{
data <- credit_data \%>\%
  first_to_lower() \%>\%
  apply_recipe(status) \%>\%
  prep(retain = TRUE) \%>\%
  juice()

rfe <- apply_rfe(data, status, c(5, 10, 20))
}
