% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_bad_rate}
\alias{calculate_bad_rate}
\title{Calculate bad rate}
\usage{
calculate_bad_rate(df, performance, top_level = "1", ...)
}
\arguments{
\item{df}{A data drame}

\item{performance}{Performance variable}

\item{top_level}{Top level of the performance variable. Defaults to 1}

\item{...}{Additional grouping variables for bad rate calculation}
}
\description{
This function calculates bad rates for any number of grouping variables.
}
\examples{
data <- recipes::credit_data \%>\%
  first_to_lower()

data \%>\%
  calculate_bad_rate(status, top_level = "bad")

data \%>\%
  mutate(status = ifelse(status == "bad", "Npl", "Pl")) \%>\%
  calculate_bad_rate(status, "Npl")

data \%>\%
  calculate_bad_rate(performance = status, top_level = "bad", marital)
}
