% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_correlation}
\alias{calculate_correlation}
\title{Calculate tidy correlation}
\usage{
calculate_correlation(df, cutoff = 0, method = "spearman",
  use = "pairwise.complete.obs", dedup = TRUE)
}
\arguments{
\item{df}{A a data frame}

\item{cutoff}{Include correlation higher then a threshold. Defaults to 0 - all variables are included}

\item{method}{Which correlation should be computed. Defaults to "spearman"}

\item{use}{Which method for computing correlation in presence of missing values. Defaults to "pairwise.complete.obs"}

\item{dedup}{Should all rows of the resulting table be deduplicated? Defaults to TRUE}
}
\description{
This function calculates correlation among all numerical attributes in a tidy format.
}
\examples{
calculate_correlation(recipes::credit_data)
}
