% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_stats_numeric}
\alias{calculate_stats_numeric}
\title{Calculate statistics of numerical attributes}
\usage{
calculate_stats_numeric(df)
}
\arguments{
\item{df}{A data frame}
}
\description{
This function calculates statistics for numerical attributes. All numerical variables
are selected by the function automatically. If you want to apply it so specific columns only
you need to specify that prior or after the function is applied. If you want to caculate
the share in a group by manner you can use the nest-mutate-map combination.
See examples for an illustration.
}
\examples{
recipes::credit_data \%>\%
  calculate_stats_numeric()

recipes::credit_data \%>\%
  first_to_lower() \%>\%
  nest(-marital) \%>\%
  mutate(stats = purrr::map(data, ~calculate_stats_numeric(.x))) \%>\%
  select(marital, stats) \%>\%
  unnest()
}
