% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak.R
\name{cap_at_percentile}
\alias{cap_at_percentile}
\title{Cap numeric values at selected percentiles}
\usage{
cap_at_percentile(x, floor = 0.025, roof = 0.975)
}
\arguments{
\item{x}{A vector or data frame}

\item{floor}{Bottom percentile. Defaults to 0.025}

\item{roof}{Top percentile. Defaults to 0.975}
}
\description{
This function capps the lowest and highest values of a numeric vector at specified percentiles.
It can be used both with numeric vectors or data frames with mutate or map_at.
}
\examples{
x <- seq(1, 100, 1)
cap_at_percentile(x)

data <- data_frame(x = seq(1, 100, 1))
data \%>\%
  mutate(y = cap_at_percentile(x))
}
