% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak.R
\name{cap_between}
\alias{cap_between}
\title{Cap numeric values between two values}
\usage{
cap_between(x, floor = NA, roof = NA)
}
\arguments{
\item{x}{A vector or data frame column}

\item{floor}{A bottom number. Defaults to NA}

\item{roof}{A top number. Defaults to NA}
}
\description{
This function capps the lowest and highest values of a numeric vector between specified values.
It can be used both with numeric vectors or data frames with mutate or map_at. If no floor and
roof values are provided the function will return exactly the same result as input values.
}
\examples{
x <- seq(1, 100, 1)
cap_between(x, 40, 60)

data <- data_frame(x = seq(1, 100, 1))

data \%<>\%
  mutate(y = cap_between(x, 40, 60))
}
