% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{find_interactions}
\alias{find_interactions}
\title{Find meaningful variable interactions}
\usage{
find_interactions(df, target, folds = 5, repeats = 1, upsample = "no")
}
\arguments{
\item{df}{A data frame}

\item{target}{A target variable}

\item{folds}{Specify the number of folds in cross-validation. Defaults to 5}

\item{repeats}{Specify the number of times the fitting process should be repeated. Defaults to 1}

\item{upsample}{Should the minority class be upsampled during resampling? Defaults to "no"}
}
\description{
This function calculates a set of potentially meaningful side-effects. Second level interactions are calculated
through estimating and tuning two models with repeated cross-validation: MARS and Lasso regression.
Data preprocessing happens automatically through applying the BP2 recipe blueprint.
}
\examples{
data <- credit_data \%>\%
  first_to_lower()

int <- find_interactions(data, status, upsample = "yes")
}
