% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak.R
\name{format_my_table}
\alias{format_my_table}
\title{Format a knitr table nicely}
\usage{
format_my_table(df, width = FALSE, font_size = 12, scroll_box = NA)
}
\arguments{
\item{df}{A data frame}

\item{width}{Should the table have full-page width? Defaults to FALSE}

\item{font_size}{What font size should be used? Defaults to 12}

\item{scroll_box}{Should the table be enframed in a scroll-box? Defaults to
NA. This option is very usefull when dealing with long tables. Must be used
as character in the following format "600px"}
}
\description{
This function creates nicelycor formatted tables in R Markdown documents. It is
designed to work with data formatting functions from the "formattable"
package. Remember that columns formating must be applied before calling the
format_my_table().
}
\examples{
credit_data \%>\%
  first_to_lower() \%>\%
  calculate_share(job) \%>\%
  format_my_table()

credit_data \%>\%
  first_to_lower() \%>\%
  calculate_share(job) \%>\%
  mutate(
     share   = formattable::percent(share, 2),
     n_group = formattable::color_tile("white", "orange")(n_group)
  ) \%>\%
  format_my_table()
}
