% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_bars}
\alias{plot_bars}
\title{Plot bar-plots of numerical variables}
\usage{
plot_bars(df, x, type_x = "num", fill = NULL, facet = NULL,
  binwidth = NULL, position = "dodge", angle = 0, title = TRUE,
  subtitle = NULL, caption = NULL, lab_x = "Value range",
  lab_y = "Count", legend = TRUE, vline = c(NaN), alpha = 1,
  quantile_low = 0.025, quantile_high = 0.975,
  palette = "cartography", theme_type = "ipsum")
}
\arguments{
\item{df}{A data frame}

\item{x}{A numeric/ categorical variable for which the bar graph is to be plotted}

\item{type_x}{Character identifier for type of the variable x defined above: "num" for numeric (plots histogram) and "char" for character (plots bar chart). Defauls to "num"}

\item{fill}{Select an additional grouping variable to be used for plotting. Defaults to NULL}

\item{facet}{Select an additional faceting variable to create facets. Defaults to NULL}

\item{binwidth}{Select binwidth, defaults to NULL and let's ggplot select the optimal binwidth}

\item{position}{Select the position of the barplot from: For numeric variables : "stack" (default), "dodge" or "fill". For categorical variables : "normal", "stacked"}

\item{angle}{Select the rotation angle for the x axis labels. Defaults to 0}

\item{title}{Should the plot title appear automatically. Defaults to TRUE}

\item{subtitle}{Text that is displayed on the subtitle. Defaults to NULL}

\item{caption}{Text that is displayed on the caption. Defaults to NULL}

\item{lab_x}{Text that is displayed on the x axis. Defaults to "Level"}

\item{lab_y}{Text that is displayed on the y axis. Defaults to "Value range"}

\item{legend}{Should the plot legend appear automatically. Defaults to TRUE}

\item{vline}{Should any horizontal lines be added to the plot. Defaults to c(NaN)}

\item{alpha}{Select plot fill transparency. Defaults to 1}

\item{quantile_low}{Select lower percentile for outliers exclusion. Defaults to 2.5\%}

\item{quantile_high}{Select upper percentile for outliers exclusion. Defaults to 97.5\%}

\item{palette}{Select a color palette from colors available in the select_palette function}

\item{theme_type}{Select a theme type from themes available in the aider_theme function}
}
\description{
This function creates nicely formatted, standardised bar-plots.
}
\examples{

data <- credit_data \%>\%
 first_to_lower()

data \%>\%
 plot_bars(x = income,
           type_x = "num",
           fill = marital,
           facet = job)
data \%>\%
 plot_bars(x = income,
           type_x = "num",
           fill = marital,
           facet = job,
           position = "stack",
           binwidth = 50,
           vline = 45,
           angle = 45,
           alpha = .7,
           palette = "berlin")

data \%>\%
  plot_bars(x = job, type_x= "char", position = "normal")

data \%>\%
  plot_bars(x = job, type_x= "char", position = "stacked", fill = status)

}
