% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_bars}
\alias{plot_bars}
\title{Plot bar-plots of numerical variables}
\usage{
plot_bars(df, x, y = NULL, y_prop = FALSE, x_type = "num",
  fill = NULL, facet = NULL, binwidth = NULL, position = "stack",
  stat = "count", angle = 0, title = TRUE, subtitle = NULL,
  caption = NULL, lab_x = "Value range", lab_y = "Proportion",
  legend = TRUE, vline = c(NaN), alpha = 1, fct_order = FALSE,
  quantile_low = 0.025, quantile_high = 0.975,
  palette = "cartography", theme_type = "grey")
}
\arguments{
\item{df}{A data frame}

\item{x}{A numeric/ categorical variable for which the bar graph is to be plotted}

\item{y}{A numeric variable which contains summarised y values, used only with stat ="identity"}

\item{y_prop}{A logical variable to choose between counts/proportion on y axis, Defaults to FALSE (proportion)}

\item{x_type}{Character identifier for type of the variable x defined above: "num" for numeric (plots histogram) and "char" for character (plots bar chart). Defauls to "num"}

\item{fill}{Select an additional grouping variable to be used for plotting. Defaults to NULL}

\item{facet}{Select an additional faceting variable to create facets. Defaults to NULL}

\item{binwidth}{Select binwidth, defaults to NULL and let's ggplot select the optimal binwidth}

\item{position}{Select the position of the barplot from: For numeric variables : "stack" (default), "dodge" or "fill".}

\item{stat}{Character identifier for whether the data is already grouped ("identity") or if the function needs to aggregate data at the level of x ("count")}

\item{angle}{Select the rotation angle for the x axis labels. Defaults to 0}

\item{title}{Should the plot title appear automatically. Defaults to TRUE}

\item{subtitle}{Text that is displayed on the subtitle. Defaults to NULL}

\item{caption}{Text that is displayed on the caption. Defaults to NULL}

\item{lab_x}{Text that is displayed on the x axis. Defaults to "Level"}

\item{lab_y}{Text that is displayed on the y axis. Defaults to "Value range"}

\item{legend}{Should the plot legend appear automatically. Defaults to TRUE}

\item{vline}{Should any horizontal lines be added to the plot. Defaults to c(NaN)}

\item{alpha}{Select plot fill transparency. Defaults to 1}

\item{fct_order}{Should the factors be reordered by their frequency? Defaults to FALSE}

\item{quantile_low}{Select lower percentile for outliers exclusion. Defaults to 2.5\%}

\item{quantile_high}{Select upper percentile for outliers exclusion. Defaults to 97.5\%}

\item{palette}{Select a color palette from colors available in the select_palette function or provide your own as a nammed vector}

\item{theme_type}{Select a theme type from themes available in the aider_theme function}
}
\description{
This function creates nicely formatted, standardised bar-plots.
}
\examples{
data <- recipes::credit_data \%>\%
 first_to_lower()

df_sum <- data \%>\%
 group_by(marital) \%>\%
 summarise(mean_inc = mean(income, na.rm = TRUE))

data \%>\%
 plot_bars(x = income,
           x_type = "num",
           fill = marital,
           facet = job)
data \%>\%
 plot_bars(x = income,
           x_type = "num",
           fill = marital,
           facet = job,
           position = "stack",
           binwidth = 50,
           vline = 45,
           angle = 45,
           alpha = .7,
           palette = "berlin")

data \%>\%
 plot_bars(x = job,
          x_type = "char",
          y_prop = FALSE) # for generating counts

data \%>\%
 plot_bars(x = job,
          x_type = "char",
          position = "dodge",
          fill = marital,
          facet = status)

data \%>\%
 plot_bars(x = job,
           x_type = "char",
           y_prop = TRUE,
           position = "fill",
           fill = marital,
           facet = status)  # for generating proportions

df_sum \%>\%
 plot_bars(x = marital,
           y = mean_inc,
           x_type = "char",
           stat ="identity")
}
