% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Plot box-plots of numerical variables}
\usage{
plot_boxplot(df, x, y, fill = NULL, facet = c(" "), ticks = 10,
  angle = 0, title = TRUE, legend = TRUE, vline = c(Inf), alpha = 0.7,
  quantile_low = 0.025, quantile_high = 0.975, pallete = "viridis")
}
\arguments{
\item{df}{A data frame}

\item{x}{A categorical variable for the x axis groupings}

\item{y}{A numerical variable for the y axis levels}

\item{fill}{Select an additional grouping variable to be used for plotting. Defaults to NULL}

\item{facet}{Select an additional faceting variable to create facets. Defaults to c(" ")}

\item{ticks}{Select the number of ticks on the y axis. Defaults to 10}

\item{angle}{Select the rotation angle for the x axis labels. Defaults to 0}

\item{title}{Should the plot title appear automatically. Defaults to TRUE}

\item{legend}{Should the plot legend appear automatically. Defaults to TRUE}

\item{vline}{Should any horizontal lines be added to the plot. Defaults to c(Inf)}

\item{alpha}{Select plot fill transparency. Defaults to .7}

\item{quantile_low}{Select lower percentile for outliers exclusion. Defaults to 2.5\%}

\item{quantile_high}{Select upper percentile for outliers exclusion. Defaults to 97.5\%}

\item{pallete}{Select a color pallete. Options are: inferno, magma, plasma and viridis. Defaults to viridis}
}
\description{
This function creates nicely formatted, standardised box-plots.
}
\examples{
data <- credit_data \%>\%
  first_to_lower()

data \%>\%
  plot_boxplot(x = marital,
               y = time)

data \%>\%
  plot_boxplot(x = marital,
               y = time,
               fill = marital)

data \%>\%
  plot_boxplot(x = marital,
               y = time,
               fill = marital,
               facet = job)

data \%>\%
  plot_boxplot(x = marital,
               y = time,
               fill = marital,
               facet = job,
               ticks = 5,
               vline = 45,
               angle = 45,
               alpha = .7)
}
