% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{train_model}
\alias{train_model}
\title{Build various predictive models}
\usage{
train_model(df, target, type = "classification", models = c("en", "rf"),
  folds = 5, repeats = 1, upsample = "no")
}
\arguments{
\item{df}{A data frame}

\item{target}{A target variable}

\item{type}{Specify the modelling task. Possible options are: "classification" (default) and "regression"}

\item{models}{Specify type of models to train. Possibile options are: "en" (Elastic-Net) and "rf" (Random Forest) as default, as well as "svm" (Support Vector Machines) and "xgb" (XgBoost)}

\item{folds}{Specify the number of folds in cross-validation. Defaults to 5}

\item{repeats}{Specify the number of times the fitting process should be repeated. Defaults to 1}

\item{upsample}{Should the minority class be upsampled during resampling? Defaults to "no"}
}
\description{
This function specifies different predictive models on the dataset. Data preprocessing happens automatically through applying the BP2 recipe blueprint.
}
\examples{
data <- credit_data \%>\%
  first_to_lower()

models <- train_model(data, status, upsample)
}
